fix crashes at shutdown on Ruby 1.9.2
Ruby 1.9.2 unconditionally destroys all the addresses
registered with rb_gc_register_address -- as well as
the global list in which they are maintained. This
leads to crashes when the Address_Registration_Guard
instances try to call rb_gc_unregister_address.

This commit disables the calls to rb_gc_unregister_address
if Ruby is shutting down. It does so by adding an at_exit
handler that globally disables all the guard destructors.