Update Hash::Iterator to work on Ruby 2.0
The Iterator was previously using internal details of the st hash
library, but Ruby 2.0 made some significant changes to the structure
that broke our code. I've rewritten the code to use public-API level
code instead and should work fine across all Rubies.

What I changed the logic to is to take the list of keys on the Hash,
which is just a straight Array, which is then trivial to keep an index
"pointer" of the current slot the Iterator is on. Then I let the Proxy
class fetch the value as needed according to the current key.