Fix complication loop error. Fixes #61
Fixing #59 created this complication error but didn't notice it until
updating to the latest Xcode clang. The problem was that other includes
pulled in Exception which itself pulls in Module_impl which tries to use
Exception before it's defined. So, the one case that Exception is used I
simply replaced with a straight std::runtime_error, as it's something
that's wrong at definition.