Support both static and shared library Ruby installs
In general, this is as simple as removing the shared-only check, but it
requires the removal of Rice::VM from the main build. There are two
reasons for this:

1) It doesn't make a lot of sense to compile in a tool that lets one
start up a Ruby interpreter inside of a Ruby interpreter, and that would
probably explode spectacularly anyway due to Ruby's global state.

2) Rice::VM is the last location of some global static variables in the
library itself, which could cause weird problems, but with point 1) we
really don't need this.

If some people are using Rice::VM, I plan on making it available as a
seperate library or just the source files themselves that can easily be
added into a project manually.