Generate the prefix on call instead of from --prefix
There's an oddity with bundler installing to system-wide gems where
bundler downloads and builds gems in a temp directory before moving them
into their final place. This was creating a bad RICE_PREFIX and causing
issues with compiling Rice gems. Looking at this, we don't need to use
prefix, as the prefix is easily calculated on demand.

Fixes #86