Re-enable the static-ruby check
After a long discussion and a lot of work in #88, we discovered that
there are two issues at play here that are out of our control.

1) Heroku runs it's own custom static-link Ruby which we can't get
working with Rice.

2) Ruby 2.2.2 and on has some serious issues with static linking in
general and there doesn't seem to be any pressure to fix this.

With that, I have to re-enable the static Ruby check so people know when
they are trying to run Rice on a platform that won't support it.