Fix compliation issues on g++ regarding RARRAY_PTR
This is also a good change in general as these macros are not
recommended for use. Switching to MRI C-API methods improves future
compatibility.

The tricky change was Rice::Array::to_c_array, as that was just
returning the raw pointer to Ruby's own data. This is getting more
dangerous as time goes on and no guarentee that this will continue to
work (ignoring compilation errors) in the future.