Remove Rice::Array::to_c_array
This was previously using RARRAY_PTR, which no longer compiles nicely in
some versions of g++ and also exposes internals of Ruby. Ruby is working
on hiding more of these internals, so we are dropping this method
entirely. Rewrote Rice::Object::vcall to do it's own building of a
c-array for rb_funcall3.