Re-enable Enum#hash and tests
The issue that we ran into here seems to have been fixed. Also, it turns
out that this mean Enum had a default #hash implementation which means
they won't work well in Hash tables. This led to a bug where the `to_s`
method was always returning `INVALID` because Rice couldn't find the
name of the enum entry in question.

Added a new test to ensure our `to_s` worked when asking C++ for an Enum
value.

Fixes #94