Formalize the concept of method vs function, where a method is associated with an instance of a Ruby/C++ object while a function is not. If a method is implemented via a function, lambda or static member then the first argument will be SELF - which can either be the Ruby object itself or a the C++ object it wraps depending on the type of the argument (VALUE vs MyClass).
The reason for this change is to enable using lambdas for define_method and define_global_function. Handed just a lambda, it is impossible to determine if it should be associated with a specific object or not, so instead we make the user tell us via define_method or define_function.

A side benefit of this change is that it removes the special code to handle Constructors and Enums (which both requires a SELF parameter and are thus methods), thereby removing the dependency from Native_Function to Object/Class.

This commit also renames Wrapped_Function to Native_Function to avoid confusion with the new Wrapper Class which is a wrapper around a raw C++ pointer/value.