Drop this section
It's pretty old and I'm not sure how useful it is to keep around at this
time. It would need to be updated to also mention pybind11. If people
are confused and start asking questions we can put this section back.

Oh, and a table of contents, that's nice to have.