It turns out immediate Type verification when calling define_method, define_attr, define_enum, define_iter doesn't work on large C++ code bases because of C++ forwarding support.
Therefore instead of immediately failing, this commit updates the TypeRegistry to track undefined types. Then when figure_type is called for the first time, which will happen when Ruby calls a C++ function that returns an object, the TypeRegistry will check the undefined types. If any still remain undefined, because they are not in the list of defined types, it will throw will throw an exception.

So if an exception is thrown it should hopefully happen almost immediately at program startup and not at some random point in the future when a method with an undefined type is called.